ax25ipd: Make LOGx() macros bulletproof.
The LOGL1..LOGL4 macros were defined like:

Now consider an invocation like:

	if (condition)
		LOGL1(...);
	else
		something_different();

CPP will expand this like:

	if (condition)
		if (loglevel>0)
			syslog(LOG_DAEMON | LOG_WARNING, ...);
		else
			something_different();

That is the else would now be considered associated with the wrong if.

Macro arguments may also not have been evaluated on every invocation
making the use of these function-like looking macros not function like.

For example:

	LOGL1("%d", i++);

would be expanded to

	if (loglevel>0)
		syslog(LOG_DAEMON | LOG_WARNING, i++);

That is depending on the value of loglevel i++ would only be incremented if
syslog was actually being called.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>