Call: Handle select(2) returning EINTR.
This may happen when call is running in curses mode, and the terminal
window is resized which results in a SIGWINCH sent.  Ncurses as normally
configured on Linux distributions will have installed a handler.  If
as it is the most likely case, the select call will be interrupted by
the signal, select will return EINTR and call will exit.

Note that this doesn't mean it's now a good idea to resize the window.
Call will still not handle the screen size change properly and the output
will look ugly and missformatted.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>