call: Fix race condition between select(2) and signal delivery.
If a signal is delivered between the start of the main select loop and select
might wait for a long time.  Fixing this with select is not possible so
switch to pselect.

Unlike select pselect akes a struct timespec pointer as its argument so
rewrite the related infrastructure to work with struct timespec instead of
struct timeval.

For now the race did only affect the SIGQUIT delivery.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>